+++ lib/epics/client.rb
@@ -123,8 +123,12 @@ class Epics::Client
     upload(Epics::CCT, document)
   end
 
-  def STA(from, to)
+  def STA(from = nil, to = nil)
     download(Epics::STA, from, to)
+  end
+
+  def VMK(from = nil, to = nil)
+    download(Epics::VMK, from, to)
   end
 
   def C52(from, to)
@@ -159,7 +163,7 @@ class Epics::Client
     download(Epics::PTK, from, to)
   end
 
-  def HAC(from, to)
+  def HAC(from = nil, to = nil)
     download(Epics::HAC, from, to)
   end
 
@@ -202,7 +206,7 @@ class Epics::Client
   end
 
   def connection
-    @connection ||= Faraday.new(headers: {user_agent: "EPICS v#{Epics::VERSION}"}) do |faraday|
+    @connection ||= Faraday.new(headers: {user_agent: "EPICS v#{Epics::VERSION}"}, ssl: { verify: verify_ssl? }) do |faraday|
       faraday.use Epics::XMLSIG, { client: self }
       faraday.use Epics::ParseEbics, { client: self}
       # faraday.response :logger                  # log requests to STDOUT
@@ -245,4 +249,7 @@ class Epics::Client
     cipher.key = OpenSSL::PKCS5.pbkdf2_hmac_sha1(passphrase, salt, 1, cipher.key_len)
   end
 
+  def verify_ssl?
+    ENV['EPICS_VERIFY_SSL'] != 'false'
+  end
 end