+++ lib/epics/hac.rb
@@ -1,10 +1,21 @@
 class Epics::HAC < Epics::GenericRequest
   attr_accessor :from, :to
 
-  def initialize(client, from, to)
+  # By default HAC only returns data for transactions which have not yet been fetched. Therefore,
+  # most applications not not have to specify a date range, but can simply fetch the status and
+  # be done
+  def initialize(client, from = nil, to = nil)
     super(client)
     self.from = from
     self.to = to
+  end
+
+  def date_range
+    if !!from && !!to
+      { "DateRange" => { "Start" => from, "End" => to } }
+    else
+      { :content! => '' }
+    end
   end
 
   def header
@@ -23,12 +34,7 @@ class Epics::HAC < Epics::GenericRequest
         "OrderDetails" => {
           "OrderType" => "HAC",
           "OrderAttribute" => "DZHNN",
-          "StandardOrderParams" => {
-            "DateRange" => {
-              "Start" => from,
-              "End" => to
-            }
-          }
+          "StandardOrderParams" => date_range
         },
         "BankPubKeyDigests" => {
           "Authentication" => {